Implement authorization using policy

+++ app/Http/Controllers/QuestionsController.php
@@ -8,7 +8,12 @@ use Illuminate\Http\Request;
 
 class QuestionsController extends Controller
 {
-    /**
+	public function __construct()
+	{
+		$this->middleware('auth', ['except' => ['index', 'show']]);
+	}
+	
+	/**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -82,6 +87,7 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+    	$this->authorize('update', $question);
       $question->update($request->only('title', 'body'));
       
       return redirect('/questions')->with('success', 'Your question has been updated');
@@ -96,6 +102,7 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+    	$this->authorize('delelte', $question);
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'You are deleted question successfully!');
     }
